requirements.txt
# pip list --format=freeze > requirements.txt
absl-py==1.3.0
aiohttp==3.8.3
aiosignal==1.2.0
albumentations==1.3.0
anyio==3.5.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.1.0
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==4.1.0
blinker==1.4
Bottleneck==1.3.5
brotlipy==0.7.0
cachetools==4.2.2
certifi==2022.12.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.4
# conda==22.11.1
# conda-build==3.23.3
# conda-package-handling==1.9.0
contourpy==1.0.5
cryptography==38.0.1
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.2.1
entrypoints==0.4
exceptiongroup==1.0.4
executing==0.8.3
expecttest==0.1.4
fastjsonschema==2.16.2
filelock==3.6.0
flit_core==3.6.0
fonttools==4.25.0
frozenlist==1.3.3
future==0.18.2
glob2==0.7
google-auth==2.6.0
google-auth-oauthlib==0.4.4
grpcio==1.42.0
hypothesis==6.61.0
idna==3.4
imageio==2.23.0
ipykernel==6.15.2
ipython==8.7.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.1
Jinja2==2.11.3
joblib==1.2.0
json5==0.9.6
jsonschema==4.16.0
jupyter==1.0.0
jupyter_client==7.4.8
jupyter-console==6.4.4
jupyter_core==4.11.2
jupyter-server==1.23.4
jupyterlab==3.5.0
jupyterlab-pygments==0.1.2
jupyterlab-server==2.10.3
jupyterlab-widgets==1.0.0
kiwisolver==1.4.2
libarchive-c==2.9
Markdown==3.4.1
MarkupSafe==2.0.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mistune==0.8.4
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
monai==1.1.0
mpmath==1.2.1
multidict==6.0.2
munkres==1.1.4
# nb-conda==2.2.1
# nb-conda-kernels==2.3.1
# nbclassic==0.4.8
# nbclient==0.5.13
# nbconvert==6.4.4
# nbformat==5.7.0
# nest-asyncio==1.5.5
networkx==2.8.8
notebook==6.5.2
notebook_shim==0.2.2
numexpr==2.8.4
numpy==1.22.3
oauthlib==3.2.1
opencv-python>=4.5
opencv-python-headless>=4.6
packaging==22.0
pandas==1.4.4
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
pip==22.3.1
pkginfo==1.8.3
pluggy==1.0.0
prometheus-client==0.14.1
prompt-toolkit==3.0.20
protobuf # ==3.20.1
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat>=0.6 # 0.6.4(ori) , available 0.6.1, 0.6.3
pycparser==2.21
Pygments==2.11.2
PyJWT==2.4.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
PyQt5==5.12.3
PyQt5_sip==4.19.18
PyQtChart==5.12
PyQtWebEngine==5.12.1
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
python-etcd==0.4.5
pytz==2022.7
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==23.2.0
qtconsole==5.3.2
QtPy==2.2.0
qudida==0.0.4
requests==2.28.1
requests-oauthlib==1.3.0
rsa==4.7.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
scikit-image==0.19.3
scikit-learn==1.2.0
scipy==1.9.3
Send2Trash==1.8.0
setuptools==65.5.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
stack-data==0.2.0
sympy==1.11.1
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
terminado==0.17.1
testpath==0.6.0
threadpoolctl==3.1.0
tifffile==2022.10.10
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==1.13.1
torchelastic==0.2.2
torchsummary==1.5.1
torchtext==0.14.1
torchvision==0.14.1
tornado==6.2
tqdm==4.64.1
traitlets==5.7.1
types-dataclasses==0.6.6
typing_extensions==4.4.0
urllib3==1.26.13
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.58.0
Werkzeug==2.0.3
wheel==0.37.1
widgetsnbextension==3.5.2
yarl==1.8.1
